#128 refactor: 공통 훅 common으로 빼기
