--- src/page/login/password/hook/useInput.ts
@@ -1,11 +1,0 @@
-import { useState } from 'react';
-
-export const useInput = () => {
-  const [value, setValue] = useState('');
-
-  const onChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setValue(e.target.value);
-  };
-
-  return { value, onChange };
-};
